requirements.txt
python-dotenv==1.0.0
pinecone-client==3.0.2
openai==1.12.0